requirements.txt
# requirements.txt

discord.py==2.0.1
openai==0.27.0
python-dotenv==1.0.0
asyncpg==0.26.0
SQLAlchemy==1.4.46
alembic==1.11.1
